% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureParam.R
\name{FutureParam}
\alias{FutureParam}
\title{Creates a FutureParam object}
\usage{
FutureParam(catch.errors = TRUE, stop.on.error = TRUE, log = FALSE,
  threshold = "INFO", logdir = NA_character_)
}
\arguments{
\item{catch.errors}{...}

\item{stop.on.error}{...}

\item{log}{...}

\item{threshold}{...}

\item{logdir}{...}
}
\value{
A \link[BiocParallel:BiocParallelParam]{BiocParallelParam} object of class FutureParam
}
\description{
Creates a FutureParam object
}

