% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocParallel.FutureParam-package.R
\docType{package}
\name{BiocParallel.FutureParam}
\alias{BiocParallel.FutureParam}
\title{BiocParallel.FutureParam: A BiocParallelParam class using Futures}
\description{
The \pkg{BiocParallel.FutureParam} package provides \link{FutureParam},
a \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} class, for the
\pkg{BiocParallel} package that works with \emph{any} type of future.
(that is supported by Future API of the \pkg{future} package) can
be used for asynchronous (parallel/distributed) or synchronous
(sequential) processing.
}
\details{
Futures themselves are provided by the \pkg{future} package, e.g.
multicore, multisession, ad hoc cluster, and MPI cluster futures.
Additional futures are provided by other packages.
For example, BatchJobs futures are implemented by the
\pkg{future.BatchJobs} package, which expands the support for
asynchronous processing to anything that the \pkg{BatchJobs}
package supports.

To use futures with the \pkg{BiocParallel} package, load
\pkg{BiocParallel.FutureParam}, call \code{register(FutureParam())},
select the type of future you wish to use via
\code{\link[=future:plan]{future:plan()}}.
}
\examples{
library("BiocParallel.FutureParam")
register(FutureParam())
plan(multiprocess)

mu <- 1.0
sigma <- 2.0
x <- bplapply(1:3, mu = mu, sigma = sigma, function(i, mu, sigma) {
  rnorm(i, mean = mu, sd = sigma)
})
print(x)

## WORKAROUND: For some reason, 'R CMD check' on Windows will give an
## error when running this example with plan(multisession), unless we
## reset the future plan at the end.
plan(sequential)
}
