% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureParam-class.R
\name{FutureParam}
\alias{FutureParam}
\title{Creates a FutureParam object}
\usage{
FutureParam(catch.errors = TRUE, stop.on.error = TRUE, log = FALSE,
  threshold = "INFO", logdir = NA_character_, ...)
}
\arguments{
\item{catch.errors}{...}

\item{stop.on.error}{...}

\item{log}{...}

\item{threshold}{...}

\item{logdir}{...}

\item{\ldots}{...}
}
\value{
A \link[BiocParallel:BiocParallelParam]{BiocParallelParam} object of class FutureParam
}
\description{
Creates a FutureParam object
}
\examples{
library("BiocParallel.FutureParam")
register(FutureParam())
plan(multiprocess)

mu <- 1.0
sigma <- 2.0
x <- bplapply(1:3, mu = mu, sigma = sigma, function(i, mu, sigma) {
  rnorm(i, mean = mu, sd = sigma)
})
print(x)

## WORKAROUND: For some reason, 'R CMD check' on Windows will give an
## error when running this example with plan(multisession), unless we
## reset the future plan at the end.
plan(sequential)
}
